# LctMonolith/openapi-gamification.yaml
openapi: 3.0.3
info:
  title: LctMonolith Gamification API
  version: 1.0.0
  description: |
    Comprehensive REST API for gamification module (players, missions, ranks, skills, rewards, dialogue, inventory, store, analytics, notifications, auth).
    Authentication via JWT Bearer (Authorization: Bearer <token>). Admin endpoints require role=Admin.
servers:
  - url: https://localhost:5001
    description: Local HTTPS
  - url: http://localhost:5000
    description: Local HTTP
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Id: { type: string, format: uuid }
...
paths:
...
  /api/skills:
    get:
      tags: [Skills]
      security: [ { bearerAuth: [] } ]
      summary: List skills
      responses: { '200': { description: OK } }
    post:
      tags: [Skills]
      security: [ { bearerAuth: [] } ]
      summary: Create skill (Admin)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string }
              required: [title]
      responses: { '201': { description: Created } }
      x-roles: [Admin]
  /api/skills/{id}:
    get:
      tags: [Skills]
      security: [ { bearerAuth: [] } ]
      summary: Get skill
      parameters:
        - in: path
          name: id
          required: true
          schema: { $ref: '#/components/schemas/Id' }
      responses: { '200': { description: OK }, '404': { description: Not Found } }
    put:
      tags: [Skills]
      security: [ { bearerAuth: [] } ]
      summary: Update skill (Admin)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string }
              required: [title]
      responses: { '200': { description: OK }, '404': { description: Not Found } }
      x-roles: [Admin]
    delete:
      tags: [Skills]
      security: [ { bearerAuth: [] } ]
      summary: Delete skill (Admin)
      responses: { '204': { description: No Content }, '404': { description: Not Found } }
      x-roles: [Admin]
  /api/skills/player/{playerId}:
    get:
      tags: [Skills]
      security: [ { bearerAuth: [] } ]
      summary: List player skills
      parameters:
        - in: path
          name: playerId
          required: true
          schema: { $ref: '#/components/schemas/Id' }
      responses: { '200': { description: OK } }
  /api/skills/player/{playerId}/{skillId}:
    post:
      tags: [Skills]
      security: [ { bearerAuth: [] } ]
      summary: Update player skill (Admin)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                level: { type: integer }
              required: [level]
      parameters:
        - in: path
          name: playerId
          required: true
          schema: { $ref: '#/components/schemas/Id' }
        - in: path
          name: skillId
          required: true
          schema: { $ref: '#/components/schemas/Id' }
      responses: { '200': { description: OK } }
      x-roles: [Admin]
...